Move implementation of merging  untouchedBoard and userBoard to <Game />
